+++ CMakeLists.txt
@@ -115,6 +115,7 @@ file(GLOB DGL_SRC
   src/runtime/*.cc
   src/geometry/*.cc
   src/geometry/cpu/*.cc
+  src/dataloading/*.cc
 )
 
 file(GLOB_RECURSE DGL_SRC_1
